+++ doubleclick-openrtb/src/test/java/com/google/doubleclick/openrtb/CompanionTypeMapperTest.java
@@ -16,10 +16,9 @@
 
 package com.google.doubleclick.openrtb;
 
-import static org.junit.Assert.assertEquals;
+import static com.google.common.truth.Truth.assertThat;
 
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableSet;
 import com.google.openrtb.OpenRtb.BidRequest.Imp.Video.VASTCompanionType;
 import com.google.protos.adx.NetworkBid.BidRequest.Video.CompanionSlot.CreativeFormat;
 
@@ -28,12 +27,10 @@ import org.junit.Test;
 public class CompanionTypeMapperTest {
   @Test
   public void testMapper() {
-    assertEquals(
-        ImmutableSet.of(VASTCompanionType.HTML),
-        CompanionTypeMapper.toOpenRtb(ImmutableList.of(CreativeFormat.HTML_CREATIVE), null));
-    assertEquals(
-        ImmutableSet.of(CreativeFormat.HTML_CREATIVE),
-        CompanionTypeMapper.toDoubleClick(ImmutableList.of(VASTCompanionType.HTML), null));
+    assertThat(CompanionTypeMapper.toOpenRtb(ImmutableList.of(CreativeFormat.HTML_CREATIVE), null))
+        .containsExactly(VASTCompanionType.HTML);
+    assertThat(CompanionTypeMapper.toDoubleClick(ImmutableList.of(VASTCompanionType.HTML), null))
+        .containsExactly(CreativeFormat.HTML_CREATIVE);
 
     for (VASTCompanionType openrtb : VASTCompanionType.values()) {
       CompanionTypeMapper.toDoubleClick(openrtb);