+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 RELEASE NOTES
 ----------------------------------------------------------------------
 
+## Version 0.9.7, ??-09-2015
+* DoubleClick protocol v.70: Refactors part of `BidRequest`'s `Mobile`
+  into a new object `Device`; OpenRTB mapping is updated for that.
+
 ## Version 0.9.6, 22-09-2015
 * DoubleClick protocol v.69: Adds unencrypted versions of several
   fields (SSL-exclusive); the mapper doesn't support the encrypted